gestion product,stock,category front

+++ src/app/dashboard/list-stock/list-stock.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { StockService } from 'src/app/services/stock.service';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-list-stock',
   templateUrl: './list-stock.component.html',
@@ -33,10 +33,28 @@ export class ListStockComponent implements OnInit {
   }
 
 deleteStock(id:number){
-  this.stockService.deleteStock(id).subscribe(data =>
-    { console.log(data);
-      this.getStocks()
-    } )
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.stockService.deleteStock(id).subscribe(data =>
+        { console.log(data);
+          this.getStocks()
+        } )
+      Swal.fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      )
+    }
+  })
+ 
 }
 
 }
